fix(python): align tests with f64 ndarrays + expose SimBox as Box
Two issues surfacing on the Python CI matrix:

1. `__init__.py` imported `SimBox` but listed `Box` in `__all__`, so
   `import molrs; molrs.Box(...)` (what every test uses) failed with
   `AttributeError`. Alias the import: `from .molrs import SimBox as Box`.

2. Tests passed `dtype=np.float32` arrays into bindings whose pyo3
   signatures are `PyReadonlyArray1<'_, f64>` (NpF is f64 now; molrs-core
   standardised on f64). PyO3 rejected the f32 ndarrays with the cryptic
   "ndarray object is not an instance of ndarray" message. Update every
   test file to use `np.float64`.

+++ molrs-python/tests/conftest.py
@@ -12,7 +12,7 @@ def cubic_box():
 @pytest.fixture
 def ortho_box():
     """A 5x10x15 orthorhombic box."""
-    return molrs.Box.ortho(np.array([5.0, 10.0, 15.0], dtype=np.float32))
+    return molrs.Box.ortho(np.array([5.0, 10.0, 15.0], dtype=np.float64))
 
 
 @pytest.fixture
@@ -26,5 +26,5 @@ def sample_points():
             [0.1, 0.1, 0.1],
             [9.9, 9.9, 9.9],
         ],
-        dtype=np.float32,
+        dtype=np.float64,
     )
